extbldr/manifest: Define manifest structure
The manifest data structure and yaml format has been defined.
Each package/repo has a single manifest which defines multiple
subpackages. Each subpackage corresponds to an SRPM to be built.

The upstream source, modified spec file and incremental patches
are also specified.
The targets(archs) to be built for are also specified. The mock
configuration for each such target is autogenerated from the target
spec in the yaml.