lemurbldr: Don't specify sources and spec in manifest
Right now sources(patches) are specified both in manifest and spec file.
Avoid this repeatition, don't specify them in the manifest.
The tool just assumes everything in sources subdir in the repo needs to
be copied over to the rpmbuild environment.

Same with spec file, it is no longer specified in the manifest. It'll be
placed in a specs subdir in the repo.

In case the repo builds mutliple SRPMs(packages), each package will have
its own subdirectory in the repo named with the same name as the package
in the manifest. A boolean flag subdir for the package has been added in
the manifest to indicate that this package is within a subdirectory.

Also fixed all the manifests in testData to honor this new syntax.

Refactored the code that copies/moves files to use Glob instead of
ReadDir and regex matches.

For tests, use require from the testify instead of assert to not continue
after first assert failure. Split TestCreateSrpm to TestCreateSrpmFromSrpm
and TestCreateSrpmFromTarball.