lemurbldr: Pick repo URLs from configuration
Addressing Ken's comments to not hardcode dnf repo specific stuff
like URLs in the manifest in each git repo.
Instead the manifest just tells the name of the repo and version,
the URL is now derived from this and a configuration file.

The configuration file specified a go template string to derive the
baseURL from the host, version and arch. version comes from manifest.
host is viper configuration and arch is a commandline arg.

Other cleanups:
1. Moved config files to their own directory.
2. Make the tests uses internal Arista dnf repos instead of public
   mirrors.
3. Make the go test workflow run on self-hosted runner as we talk to
   internal mirrors.