changes for building i686 rpms with alma vault mirrors
The original design was to build i686 RPMs with x86_64 repos itself
using multilib, but dnf and mock have been hard to tame to make them do
this. So we're planning to build with native alma i686 mirrors available
in vault. To make this happen, baseURL derivation has to insert `i686`
in repo URL. It's been changed to do this by default unless explicitly
asked to add the baseArch of `x86_64` using use-base-arch in the repo
configuration in eext.yaml.

I've modified the default dnfrepoconfig.yaml to use alma vault mirrors
in the base URL template. The version can be specified specifically
like `9.1` or as say `default` or `latest` in case we want to build with
new BuildRequires repo when new version comes out and we change
`default` in dnfrepoconfig.yaml to point to a new version without any
changes to eext.yaml.