Don't run installdeps as a separate stage for target i686
We run installdeps as a separate stage from build to make debugging easier.
However, I see that installdeps doesn't handle %ifarch conditionals for
BuildRequires and Requires correctly.
I see that `%ifarch x86_64` evaluates to true when run with --target i686,
pulling in the wrong dependencies. So don't run it as a separate stage when
target is i686.