repoconfig: Use repo bundles
Pick repos as a bundle in eext.yaml instead of specifying each sub-repo.
I've added a bundle named el9 pointing to alma-vault cache in artifactory,
and another one, epel9, pointing to centos-9-epel cache in artifactory.

We no longer need to specify sub-repos like  BaseOS, CRB, AppStream etc
in eext.yaml, instead we just say el9. All the three  repos above are in
the bundle, and enabled by default. devel is another sub-repo which is
in the el9 bundle, but not enabled by default. This needs to be enabled
by extra lines in eext.yaml. This is not enabled by default to make dnf
faster as most packages don't need this for BuildRequires.

Fixes: BUG838343