Fix how rpm release macros are defined
Don't make it mandatory for release macro be specified in manifest.
If it is unspecified, construct it from commit hashes of SRC_<N>
environment variables.

If SRC_<N> env vars are unset too, don't define the release macro,
but leave it to the spec file to insert an `eng` as release.

Fixes: BUG838064