mock: Add an `eextsig` field to the distribution rpm metadata tag
This is to embed a signature of the source and buildrequires used to
build the rpms by eext. The signature is derived by concatening all the
SRC_<N> env vars with a `,` separator.

The idea here is to embed this information so that it can be used as a
substitue for content signatures by Abuild in cases where the rpm builds
produce non-idempotent content signatures.

Fixes: BUG838065