eext: Implement chained builds
Add a new attribute `local-deps` in the build spec of the manifest.
Add a new config DepsDir which is the directory containing local RPM
dependencies, default it to /RPMS so that barney build output of
dependencies can be pulled directly in to the build floor.

When building packages with local-deps, all the dependencies in DepsDir
are copied over to a sub-path in the working directory, then createrepo
is run there to generate the yum metadata, and the auto-generated mock
configuration has a new `local-deps` repo setup to point to this
directory with the dependency RPMs and yum metadata with a `file://`
URL.

Enhanced mock_cfg_test to verify that the new repo is generated when the
manifest specifies a chained build. Also added a separate test to make
sure the deps are copied over and that the yum metadata is generated.

Fixes: BUG837043