dnfconfig: Remove epel9 bundle
The epel9 repo-bundle pointed to a moving dnf repository which is bad
for reproducible eext builds.

Instead of epel9, use a new repo-bundle epel9-subset which is a
frozen/versioned subset of epel9 managed locally. It only has a few epel
packages like quilt, hiredis etc required by existing eext third-party
repos as on the date of this change.