cmd: suppress command usage on execution errors
Cobra, by default, has the odd behaviour of printing out the command
usage whenever anything in the command throws an error.

It's really strange to have the mock build fail, for example, and then
have the usage for "eext build" displayed.

This change stops printing out usage in all error cases. This includes
cases where invalid arguments are provided where, arguably, the usage
would be helpful, but in any case, the user can just pass "-h" or
"--help" to see the usage instead.