Bootstrap without docker
With this change we bootstrap the workspace creation environment using a
centos rootfs rather than pulling down a docker container. This has the
following benefits:

 - docker and golang are no longer in the dependency set which should
result in less cache churn
 - we no longer need a finalizer to set up yum repos --- creating these
files in their own snapshot and overlaying it makes it much more
cache-friendly and avoids us needing to mount the whole workspace in a
tmpfs in order to let the finalizer run
 - we can check the content hash of the thing we download (in theory we
trust quay.io to not change their docker image tags, but we have no way
of checking this)
 - The artifactory URL can now be overridden via cluster config -- although
the default value of artifactory.infra.corp.arista.io is preserved

It does add a bit of complexity that we start with a centos container to
bootstrap an alma container. However I don't think this is really an
issue as the content that we are installing comes from our internal alma
repositories. The only thing we are using centos for is to run dnf and
resolve/install dependencies into our alma images.