Adding support el9-snapshot in dnfconfig
Adding el9-snapshot repo-bundle to help the eext team run experiments
to see if builds are compatible with newer Buildrequires. These point
to a snapshot captured at the given date of the el9 upstream repo of
alma-linux (x86_64 and aarch64 rpms) and alma-vault (i686 rpms).

Please note that we haven't specified any default snapshot value for the
repo-bundle. This is because users should use el9 repo-bundle whenever
possible. Only in the case of the package requiring the latest release
(which hasn't been vaulted yet), the eext team will snapshot the
el9-unsafe repo and provide the snapshot version, which the user should
use in their repo-bundle.