Priority for repo-bundles and repos
Some rpms can be distributed in multiple repos, hence we need to specify
priority to the repo for eext to decide where to pick dependencies from.

Priority 1 (highest) is given to the local deps available for eext to
check before checking upstream repos.

The priority should be defined at a repo-bundle level in dnfconfig.yaml,
with a priorty > 1.
Or within the package specific eext.yaml file at a repo specific level.
This is an optional flag, the default priority is assigned by the
repo-bundle.