Avoid using .%go/static
The go generator would pull in the current repo's commit hash into the
snapshot hash. Instead, we'll running go build ourselves on a floor with
the go toolchain and modules already present. We also map in the subset
of sources needed to build the eext static binary.

.%go/modules would have sufficed if not for the fact that it also pulls
in all sources to do go mod tidy. So we just enable the network and
forego any benefit from the go generator module cache.