Support for git-worktree
We have currently enabled support to use an upstream git repo as a
source for eext. However there are cases where developers have an
internal repo (forked from upstream) that they want to use as the
source. We hence add support for 'git-worktree'.
Developers can add an 'eext.yaml' file in their repo (no need to create
an eext specific repo), along with the .spec file to build the package.
eext will take care of creating the source tarball from the git repo,
and generate the necessary (S)RPMs.