explicitly pass config path when reading dnf config
This is the first patch in series to reduce the reliance on the global,
implic viper object. Making path an argument to the loading function
this patch enables us to do a simple dependency injection in tests
instead of having to set (and then reset via defer) the value in the
global viper object.