mark some tests as "containerized" and make barney run them
This commit brings a disctinction between unit/code tests and a
production-like tests. The latter do destructive changes to the system
they're being run on and they should be run in container, thus the
"containerized" tag.

Barney already run all the tests, including the destructive ones, so
this patch also makes sure that barney runs the newly marked tests.

Overall this patch enables developers to run the normal unit and
integration tests without having to reach to Barney to iterate.

The follow-up patch would be to mock all the system tools called by the
tests and mocking them inside the code, creating the non-destructive
alternative set that could be run on any machine.