simplify the RPM specification file
The rpms-* files all started with lines beginning with '#' which would
suggest a unix-script-style comment, but when the '#' was applied as an
inline comment, the install-rpms script treated that as part of the
package name, breaking it. Having those comments also meant that they
have to be stripped when processing.

This patch moves the content of those comments to a standalone README,
simplifying the sytnax of the file so it can be used as-is in the
install-rpms script.