fix and simplify install-rpms
The install-rpms accepted two non-optional "options" that ended up being
used the same way. There was some logic in that bash script that checked
and processed those options, and as it's typical of bash code, it had a
bug, where if the script was supplied with only one of the two
"options", the value of non-existent $extra_rpms_file, because it was
enclosed in double quotes, would be treated as an actual object (an
empty string), thus making the following code:
```
for file in "$common_rpms_file" "$extra_rpms_file"; do
(...)
mapfile -t tmp_array < <(awk '!/^#/' "$file")
```
iterate over an empty "$file" blocking awk.

Because the only thing that this script did was to install RPMs
specified by text files, this patch changes it to do exactly that
without any error-prone bash logic.