Upgrade dependencies to latest versions
With release of Alma 9.5, we want to upgrade the eext package
dependencies to the latest vaulted versions of the corresponding rpms.
This ensures eext packages are using the latest rpms to build.
We update the el9 repo to point to 9.4 (vaulted), and create an epel
repo snapshot containing the updated versions of the rpms.