Add `list-unverified-sources` root command and SHA-256 hash support for unverified upstream sources
This new root command is part of the `stest` and is designed to list all upstream sources with the
`skip-check` flag set to `true` for a specified package. The output is written to a file at the
following location:  `/dest/code.arista.io/eos/eext/{package}.unverifiedSources.json`.
This file will be included in the snapshot build created by Barney, ensuring better tracking of
unverified sources.

Unverified upstream sources (those with `skip-check: true`) are inherently risky as they lack a
verification method. Any changes to these sources could go undetected. To address this, we now
calculate and store the SHA-256 hash of unverified sources. This hash is added to the `eext.yaml`
file under the `src-sha256-hash` field. During the `create-srpm` command, the hash in `eext.yaml`
will be compared with the hash of the downloaded sources.