add mockable executor
This patch introduces a layer over os.exec that makes it possible to
write testable code that doesn't run actual commands. The common
executor interface is acompanied by two implementations, on that runs
the commands normally (as the previous code did), and a "Dry Run"
executor that keeps track of what would be executed. This enables us to
have a "dry-run" mode that instead of executing the commands writes down
what would be run, making it easy to run the same commands by hand when
debugging.

See the unit tests for more details.