Don't run build-prep by default for create-srpm
Deprecated the "--skip-build-prep" option, making it a no-op, and
build-prep is now skipped by default. The option "--do-build-prep" has
been added for users wanting previous behavior. We cannot remove
the option altogether because there're packages which make use
of this flag via eextgen.

The motivation is that users have been stumped by %prep failures because
of missing dependencies in the their environment. It's not obvious that
--skip-build-prep will workaround the problem.